# dist/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
...
			"ChallengeMissionResponse": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"userId": {
						"type": "number",
						"format": "double"
					},
					"missionId": {
						"type": "number",
						"format": "double"
					},
					"status": {
						"type": "boolean"
					},
					"dueDate": {
						"type": "string",
						"nullable": true
					},
					"createdAt": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"userId",
					"missionId",
					"status",
					"dueDate",
					"createdAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ApiResponse_ChallengeMissionResponse_": {
				"properties": {
					"resultType": {
						"type": "string",
						"enum": [
							"SUCCESS"
						],
						"nullable": false
					},
					"error": {
						"type": "number",
						"enum": [
							null
						],
						"nullable": true
					},
					"data": {
						"$ref": "#/components/schemas/ChallengeMissionResponse"
					}
				},
				"required": [
					"resultType",
					"error",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
			"MissingMissionIdData": {
				"properties": {
					"missionId": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"missionId"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ErrorResponse_MissingMissionIdData_": {
				"properties": {
					"resultType": {
						"type": "string",
						"enum": [
							"FAILED"
						],
						"nullable": false
					},
					"error": {
						"properties": {
							"data": {
								"$ref": "#/components/schemas/MissingMissionIdData"
							},
							"message": {
								"type": "string"
							},
							"statusCode": {
								"type": "number",
								"format": "double"
							},
							"errorCode": {
								"type": "string"
							}
						},
						"required": [
							"data",
							"message",
							"statusCode",
							"errorCode"
						],
						"type": "object"
					},
					"data": {
						"type": "number",
						"enum": [
							null
						],
						"nullable": true
					}
				},
				"required": [
					"resultType",
					"error",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
			"MissionNotFoundData": {
				"properties": {
					"userId": {
						"type": "number",
						"format": "double"
					},
					"missionId": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"userId",
					"missionId"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ErrorResponse_MissionNotFoundData_": {
				"properties": {
					"resultType": {
						"type": "string",
						"enum": [
							"FAILED"
						],
						"nullable": false
					},
					"error": {
						"properties": {
							"data": {
								"$ref": "#/components/schemas/MissionNotFoundData"
							},
							"message": {
								"type": "string"
							},
							"statusCode": {
								"type": "number",
								"format": "double"
							},
							"errorCode": {
								"type": "string"
							}
						},
						"required": [
							"data",
							"message",
							"statusCode",
							"errorCode"
						],
						"type": "object"
					},
					"data": {
						"type": "number",
						"enum": [
							null
						],
						"nullable": true
					}
				},
				"required": [
					"resultType",
					"error",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
			"AlreadyChallengingMissionData": {
				"properties": {
					"userId": {
						"type": "number",
						"format": "double"
					},
					"missionId": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"userId",
					"missionId"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ErrorResponse_AlreadyChallengingMissionData_": {
				"properties": {
					"resultType": {
						"type": "string",
						"enum": [
							"FAILED"
						],
						"nullable": false
					},
					"error": {
						"properties": {
							"data": {
								"$ref": "#/components/schemas/AlreadyChallengingMissionData"
							},
							"message": {
								"type": "string"
							},
							"statusCode": {
								"type": "number",
								"format": "double"
							},
							"errorCode": {
								"type": "string"
							}
						},
						"required": [
							"data",
							"message",
							"statusCode",
							"errorCode"
						],
						"type": "object"
					},
					"data": {
						"type": "number",
						"enum": [
							null
						],
						"nullable": true
					}
				},
				"required": [
					"resultType",
					"error",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ListUserMissionsResponse": {
				"properties": {
					"missions": {
						"items": {
							"$ref": "#/components/schemas/ChallengeMissionResponse"
						},
						"type": "array"
					},
					"pagination": {
						"properties": {
							"cursor": {
								"type": "number",
								"format": "double",
								"nullable": true
							}
						},
						"required": [
							"cursor"
						],
						"type": "object"
					}
				},
				"required": [
					"missions",
					"pagination"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ApiResponse_ListUserMissionsResponse_": {
				"properties": {
					"resultType": {
						"type": "string",
						"enum": [
							"SUCCESS"
						],
						"nullable": false
					},
					"error": {
						"type": "number",
						"enum": [
							null
						],
						"nullable": true
					},
					"data": {
						"$ref": "#/components/schemas/ListUserMissionsResponse"
					}
				},
				"required": [
					"resultType",
					"error",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CompleteMissionResponse": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"userId": {
						"type": "number",
						"format": "double"
					},
					"missionId": {
						"type": "number",
						"format": "double"
					},
					"status": {
						"type": "boolean"
					},
					"dueDate": {
						"type": "string",
						"nullable": true
					},
					"createdAt": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"userId",
					"missionId",
					"status",
					"dueDate",
					"createdAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ApiResponse_CompleteMissionResponse_": {
				"properties": {
					"resultType": {
						"type": "string",
						"enum": [
							"SUCCESS"
						],
						"nullable": false
					},
					"error": {
						"type": "number",
						"enum": [
							null
						],
						"nullable": true
					},
					"data": {
						"$ref": "#/components/schemas/CompleteMissionResponse"
					}
				},
				"required": [
					"resultType",
					"error",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
			"MissionNotInProgressData": {
				"properties": {
					"userId": {
						"type": "number",
						"format": "double"
					},
					"missionId": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"userId",
					"missionId"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ErrorResponse_MissionNotInProgressData_": {
				"properties": {
					"resultType": {
						"type": "string",
						"enum": [
							"FAILED"
						],
						"nullable": false
					},
					"error": {
						"properties": {
							"data": {
								"$ref": "#/components/schemas/MissionNotInProgressData"
							},
							"message": {
								"type": "string"
							},
							"statusCode": {
								"type": "number",
								"format": "double"
							},
							"errorCode": {
								"type": "string"
							}
						},
						"required": [
							"data",
							"message",
							"statusCode",
							"errorCode"
						],
						"type": "object"
					},
					"data": {
						"type": "number",
						"enum": [
							null
						],
						"nullable": true
					}
				},
				"required": [
					"resultType",
					"error",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"CreateMissionResponse": {
				"properties": {
					"missionId": {
						"type": "number",
						"format": "double"
					},
					"storeId": {
						"type": "number",
						"format": "double"
					},
					"name": {
						"type": "string"
					},
					"minPay": {
						"type": "number",
						"format": "double"
					},
					"reward": {
						"type": "number",
						"format": "double"
					},
					"missionDue": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"missionId",
					"storeId",
					"name",
					"minPay",
					"reward",
					"missionDue"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ApiResponse_CreateMissionResponse_": {
				"properties": {
					"resultType": {
						"type": "string",
						"enum": [
							"SUCCESS"
						],
						"nullable": false
					},
					"error": {
						"type": "number",
						"enum": [
							null
						],
						"nullable": true
					},
					"data": {
						"$ref": "#/components/schemas/CreateMissionResponse"
					}
				},
				"required": [
					"resultType",
					"error",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"InvalidMissionDueData": {
				"properties": {
					"missionDue": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"missionDue"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ErrorResponse_InvalidMissionDueData_": {
				"properties": {
					"resultType": {
						"type": "string",
						"enum": [
							"FAILED"
						],
						"nullable": false
					},
					"error": {
						"properties": {
							"data": {
								"$ref": "#/components/schemas/InvalidMissionDueData"
							},
							"message": {
								"type": "string"
							},
							"statusCode": {
								"type": "number",
								"format": "double"
							},
							"errorCode": {
								"type": "string"
							}
						},
						"required": [
							"data",
							"message",
							"statusCode",
							"errorCode"
						],
						"type": "object"
					},
					"data": {
						"type": "number",
						"enum": [
							null
						],
						"nullable": true
					}
				},
				"required": [
					"resultType",
					"error",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
			"AreaNotFoundFromMissionCreateData": {
				"properties": {
					"storeId": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"storeId"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ErrorResponse_AreaNotFoundFromMissionCreateData_": {
				"properties": {
					"resultType": {
						"type": "string",
						"enum": [
							"FAILED"
						],
						"nullable": false
					},
					"error": {
						"properties": {
							"data": {
								"$ref": "#/components/schemas/AreaNotFoundFromMissionCreateData"
							},
							"message": {
								"type": "string"
							},
							"statusCode": {
								"type": "number",
								"format": "double"
							},
							"errorCode": {
								"type": "string"
							}
						},
						"required": [
							"data",
							"message",
							"statusCode",
							"errorCode"
						],
						"type": "object"
					},
					"data": {
						"type": "number",
						"enum": [
							null
						],
						"nullable": true
					}
				},
				"required": [
					"resultType",
					"error",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CreateMissionRequest": {
				"properties": {
					"name": {
						"type": "string"
					},
					"minPay": {
						"type": "number",
						"format": "double"
					},
					"reward": {
						"type": "number",
						"format": "double"
					},
					"missionDue": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"name",
					"minPay",
					"reward",
					"missionDue"
				],
				"type": "object",
				"additionalProperties": false
			},
			"MissionListItem": {
				"properties": {
					"missionId": {
						"type": "number",
						"format": "double"
					},
					"storeId": {
						"type": "number",
						"format": "double"
					},
					"name": {
						"type": "string"
					},
					"minPay": {
						"type": "number",
						"format": "double"
					},
					"reward": {
						"type": "number",
						"format": "double"
					},
					"missionDue": {
						"type": "number",
						"format": "double"
					},
					"createdAt": {
						"type": "string"
					}
				},
				"required": [
					"missionId",
					"storeId",
					"name",
					"minPay",
					"reward",
					"missionDue",
					"createdAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"MissionListResponse": {
				"properties": {
					"missions": {
						"items": {
							"$ref": "#/components/schemas/MissionListItem"
						},
						"type": "array"
					},
					"pagination": {
						"properties": {
							"cursor": {
								"type": "number",
								"format": "double",
								"nullable": true
							}
						},
						"required": [
							"cursor"
						],
						"type": "object"
					}
				},
				"required": [
					"missions",
					"pagination"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ApiResponse_MissionListResponse_": {
				"properties": {
					"resultType": {
						"type": "string",
						"enum": [
							"SUCCESS"
						],
						"nullable": false
					},
					"error": {
						"type": "number",
						"enum": [
							null
						],
						"nullable": true
					},
					"data": {
						"$ref": "#/components/schemas/MissionListResponse"
					}
				},
				"required": [
					"resultType",
					"error",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "umc-10th-node.js-neon",
		"version": "1.0.0",
		"description": "��#\u0000 \u0000u\u0000m\u0000c\u0000-\u00001\u00000\u0000t\u0000h\u0000-\u0000n\u0000o\u0000d\u0000e\u0000.\u0000j\u0000s\u0000-\u0000n\u0000e\u0000o\u0000n\u0000\r\u0000 \u0000",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"paths": {
...
		"/users/missions/{missionId}/challenges": {
			"post": {
				"operationId": "HandleChallengeMission",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_ChallengeMissionResponse_"
								}
							}
						}
					},
					"400": {
						"description": "필수 입력값 누락",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse_MissingMissionIdData_"
								},
								"examples": {
									"Example 1": {
										"value": {
											"resultType": "FAILED",
											"error": {
												"errorCode": "E4001",
												"statusCode": 400,
												"message": "필수 입력값이 누락되었습니다.",
												"data": {
													"missionId": 0
												}
											},
											"data": null
										}
									}
								}
							}
						}
					},
					"404": {
						"description": "미션을 찾을 수 없음",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse_MissionNotFoundData_"
								},
								"examples": {
									"Example 1": {
										"value": {
											"resultType": "FAILED",
											"error": {
												"errorCode": "M004",
												"statusCode": 404,
												"message": "존재하지 않는 미션입니다.",
												"data": {
													"userId": 1,
													"missionId": 1
												}
											},
											"data": null
										}
									}
								}
							}
						}
					},
					"409": {
						"description": "이미 도전 중인 미션",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse_AlreadyChallengingMissionData_"
								},
								"examples": {
									"Example 1": {
										"value": {
											"resultType": "FAILED",
											"error": {
												"errorCode": "M002",
												"statusCode": 409,
												"message": "이미 도전 중인 미션입니다.",
												"data": {
													"userId": 1,
													"missionId": 1
												}
											},
											"data": null
										}
									}
								}
							}
						}
					}
				},
				"description": "미션 도전 시작",
				"summary": "사용자가 특정 미션에 도전을 시작",
				"tags": [
					"User Missions"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "missionId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/users/missions": {
			"get": {
				"operationId": "HandleListUserMissions",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_ListUserMissionsResponse_"
								}
							}
						}
					}
				},
				"description": "사용자 미션 목록 조회",
				"summary": "로그인한 사용자의 진행 중 또는 완료된 미션 목록을 조회",
				"tags": [
					"User Missions"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "cursor",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "status",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/users/missions/{missionId}/success": {
			"post": {
				"operationId": "HandleCompleteMission",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_CompleteMissionResponse_"
								}
							}
						}
					},
					"400": {
						"description": "진행 중인 미션 아님",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse_MissionNotInProgressData_"
								},
								"examples": {
									"Example 1": {
										"value": {
											"resultType": "FAILED",
											"error": {
												"errorCode": "E4001",
												"statusCode": 400,
												"message": "필수 입력값이 누락되었습니다.",
												"data": {
													"missionId": 0
												}
											},
											"data": null
										}
									},
									"Example 2": {
										"value": {
											"resultType": "FAILED",
											"error": {
												"errorCode": "M003",
												"statusCode": 400,
												"message": "진행 중의 미션이 아닙니다.",
												"data": {
													"userId": 1,
													"missionId": 1
												}
											},
											"data": null
										}
									}
								}
							}
						}
					}
				},
				"description": "미션 완료 처리",
				"summary": "진행 중인 미션을 완료 처리",
				"tags": [
					"User Missions"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "missionId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
...
		"/stores/{storeId}/missions/write": {
			"post": {
				"operationId": "HandleCreateMission",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_CreateMissionResponse_"
								}
							}
						}
					},
					"400": {
						"description": "유효하지 않은 기한",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse_InvalidMissionDueData_"
								},
								"examples": {
									"Example 1": {
										"value": {
											"resultType": "FAILED",
											"error": {
												"errorCode": "E4001",
												"statusCode": 400,
												"message": "storeId가 올바르지 않습니다.",
												"data": {
													"storeId": 0
												}
											},
											"data": null
										}
									},
									"Example 2": {
										"value": {
											"resultType": "FAILED",
											"error": {
												"errorCode": "E4001",
												"statusCode": 400,
												"message": "미션 이름은 1~20자 사이여야 합니다.",
												"data": {
													"name": ""
												}
											},
											"data": null
										}
									},
									"Example 3": {
										"value": {
											"resultType": "FAILED",
											"error": {
												"errorCode": "E4001",
												"statusCode": 400,
												"message": "최소 금액은 0보다 커야 합니다.",
												"data": {
													"minPay": 0
												}
											},
											"data": null
										}
									},
									"Example 4": {
										"value": {
											"resultType": "FAILED",
											"error": {
												"errorCode": "E4001",
												"statusCode": 400,
												"message": "보상금은 0보다 커야 합니다.",
												"data": {
													"reward": 0
												}
											},
											"data": null
										}
									},
									"Example 5": {
										"value": {
											"resultType": "FAILED",
											"error": {
												"errorCode": "E4001",
												"statusCode": 400,
												"message": "기한은 0보다 커야 합니다.",
												"data": {
													"missionDue": 0
												}
											},
											"data": null
										}
									}
								}
							}
						}
					},
					"404": {
						"description": "존재하지 않는 가게",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse_AreaNotFoundFromMissionCreateData_"
								},
								"examples": {
									"Example 1": {
										"value": {
											"resultType": "FAILED",
											"error": {
												"errorCode": "S001",
												"statusCode": 404,
												"message": "존재하지 않는 가게입니다.",
												"data": {
													"storeId": 1
												}
											},
											"data": null
										}
									}
								}
							}
						}
					}
				},
				"description": "새로운 미션 생성",
				"summary": "특정 가게에 새로운 미션을 생성",
				"tags": [
					"Missions"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "storeId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateMissionRequest"
							}
						}
					}
				}
			}
		},
		"/stores/{storeId}/missions": {
			"get": {
				"operationId": "HandleListStoreMissions",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_MissionListResponse_"
								}
							}
						}
					}
				},
				"description": "가게의 미션 목록 조회",
				"summary": "특정 가게의 모든 미션을 조회합니다. 커서 기반 페이지네이션을 사용",
				"tags": [
					"Missions"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "storeId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "cursor",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		}
	},
	"servers": [
		{
			"url": "/api/v1"
		}
	]
}